+++ src/MathOp.h
@@ -39,7 +39,7 @@ public:
     static double log_sum(const vector<double> &vec); // vector version
     static double log_replace_sum(const double& sum, const vector<double>& to_subtract, const vector<double>& to_add);
     static vector<double> combine_scores(vector<double> aic_vec);
-    static vector<vector<double>> likelihood_ratio(vector<vector<double>>& mat, double window_size);
+    static vector<vector<double>> likelihood_ratio(vector<vector<double>> &mat, int window_size);
     static double breakpoint_log_prior(int k, int m, double mu);
     static double log_n_choose_k(int n, int k);
     static double n_choose_k(int n, int k);
@@ -57,7 +57,7 @@ public:
     static double compute_omega_insert_delete(Node *node, double lambda_r, double lambda_c, double K);
     static double compute_omega_condense_split(Node *node, double lambda_s, int n_regions, bool weighted);
     static double frobenius_avg(vector<vector<int>>& mat, vector<vector<int>>& ground_truth);
-    static vector<long double> dirichlet_sample(int len, double alpha = 1.0);
+    static vector<long double> dirichlet_sample(size_t len, double alpha = 1.0);
 
 };
